requirements.txt
fastapi
uvicorn
requests
google-generativeai
PyPDF2
